Eliminate matrix inversion
Camera's view matrix is an inverse of its transformation matrix.

Previously, the view matrix is calculated as follows

    View = inv( Translation * Rotation * Scale )

which can be converted to

    View = inv( Scale ) * inv( Rotation ) * inv( Translation )

The inverse of scale matrix is calculated as follows

   [ 1/Sx 0    0    0 ]
   [ 0    1/Sy 0    0 ]
   [ 0    0    1/Sz 0 ]
   [ 0    0    0    1 ]

The inverse of translation matrix is calculated as follows

   [ 0  0  0  0 ]
   [ 0  0  0  0 ]
   [ 0  0  0  0 ]
   [ -x -y -z 1 ]

The inverse of rotation matrix is its own transpose since a rotation matrix
is an orthogonal matrix. http://en.wikipedia.org/wiki/Orthogonal_matrix